only sync kernel modules with `.sync.conf` extension

+++ .config/sys/kernel/backup-modules.sh
@@ -5,7 +5,13 @@ set -e;
 source "$XDG_CONFIG_HOME/sys/kernel/common-modules.sh";
 
 backup_files=$(list_backup_modules);
-rm --recursive $backup_files;
+
+if [ ! -z $backup_files ]; then
+  rm --recursive $backup_files;
+fi
 
 modules_file_path=$(list_modules);
-cp --recursive $modules_file_path "$BACKUP_PATH/";
+
+if [ ! -z $modules_file_path ]; then
+  cp --recursive $modules_file_path "$BACKUP_PATH/";
+fi
